% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQSmooth.R
\name{addQSmooth}
\alias{addQSmooth}
\alias{addQSmooth,SummarizedExperiment-method}
\title{Smooth Quantile Normalise a SummarizedExperiment}
\usage{
addQSmooth(x, ...)

\S4method{addQSmooth}{SummarizedExperiment}(x, assay = "counts", factor, n_w = 10000, ...)
}
\arguments{
\item{x}{a RangedSummarizedExperiment object}

\item{...}{Passed to \link[qsmooth]{qsmooth}}

\item{assay}{The assay to apply SmoothQuantileNormalisation to}

\item{factor}{The column used to group the data. Must be one of the columns
in the colData element of x}

\item{n_w}{The maximum number of weights to return in the metadata}
}
\value{
A SummarizedExperiment with additional metadata and the qsmooth assay
}
\description{
Smooth Quantile Normalise a SummarizedExperiment
}
\details{
This function adds the assay 'qsmooth' to a SummarizedExperiment or any
derivative class, choosing which existing assay to normalise.

A subsample up to 10,000 of the qsmooth weights are returned in the metadata
element of the final object.
}
\examples{
dat <- cbind(
  matrix(rnorm(1000), nrow=100, ncol=5),
  matrix(rnorm(1000, .1, .7), nrow=100, ncol=5)
 )
colnames(dat) <- c(paste0("A", 1:5), paste0("B", 1:5))
df <- DataFrame(group = rep(c("A", "B"), each = 5))
se <- SummarizedExperiment(
  assays = SimpleList(counts = dat),
  colData = df
)
se2 <- addQSmooth(se, "counts", "group")

}
