% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVoomWeights.R
\name{addVoomWeights}
\alias{addVoomWeights}
\alias{addVoomWeights,SummarizedExperiment-method}
\title{Add Voom Weights to a SummarizedExperiment}
\usage{
addVoomWeights(x, ...)

\S4method{addVoomWeights}{SummarizedExperiment}(
  x,
  assay = "counts",
  design = NULL,
  w0,
  isLogCPM = TRUE,
  ...,
  span = 0.5
)
}
\arguments{
\item{x}{A SummarizedExperiment object}

\item{...}{Passed to \link[limma]{voom} and \link[limma]{arrayWeights}}

\item{assay}{The assay to use for estimation of Voom Precision Weights}

\item{design}{The design matrix}

\item{w0}{Vector of initial weights as calculated by
\link[limma]{arrayWeights}}

\item{isLogCPM}{logical(1). Ignored if assay = "counts"}

\item{span}{See \link[limma]{voom}}
}
\value{
A SummarizedExperiment object
}
\description{
Add Voom Weights to a Summarized Experiment from counts or CPM
}
\details{
This adds an assay titled "voomweights" to the SummarizedExperiment object.
These can be calculated directly from the "counts" assay or from any other.
In particular, this is intended to be used on CPM or normalised CPM values.
Calling this directly on the counts assay will use the standard
\link[limma]{voom} approach, whilst other assays will assume CPM values
}
