% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangesAsTibble.R
\name{as_tibble.DataFrame}
\alias{as_tibble.DataFrame}
\alias{as_tibble.GenomicRanges}
\title{Convert GRanges or DFrame to a tibble}
\usage{
\method{as_tibble}{DataFrame}(x, ...)

\method{as_tibble}{GenomicRanges}(x, name = "range", rangeAsChar = TRUE, ...)
}
\arguments{
\item{x}{A Genomic Ranges or DataFrame object}

\item{...}{Not used}

\item{name}{Name of column to use for ranges. Ignored if rangeAsChar =
\code{FALSE}}

\item{rangeAsChar}{Convert the GRanges element to a character vector}
}
\value{
A \link[tibble]{tibble}
}
\description{
Convert GRanges or DataFrame objects to tibble objects
}
\details{
Quick and dirty conversion into a tibble. The ranges will be
returned as a character column called \code{range}. Seqinfo information will be
lost.
}
\examples{
gr <- GRanges("chr1:1-10")
gr$p_value <- runif(1)
as_tibble(mcols(gr))
as_tibble(gr)

}
