% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asTibble.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble.DataFrame}
\alias{as_tibble.GenomicRanges}
\title{Convert a GRanges or DataFrame to a tibble}
\usage{
\method{as_tibble}{DataFrame}(x, rangeAsChar = TRUE, ...)

\method{as_tibble}{GenomicRanges}(x, rangeAsChar = TRUE, name = "range", ...)
}
\arguments{
\item{x}{A Genomic Ranges or DataFrame object}

\item{rangeAsChar}{Convert any GRanges element to a character vector}

\item{...}{Not used}

\item{name}{Name of column to use for ranges. Ignored if rangeAsChar =
\code{FALSE}}
}
\value{
A \link[tibble]{tibble}
}
\description{
Convert GRanges or DataFrame objects to tibble objects
}
\details{
Quick and dirty conversion into a tibble. The ranges will be
returned as a character column called \code{range}. Seqinfo information will be
lost.

Any Compressed/SimpleList columns will be coerced to S3 list columns

Defined as an S3 method for compatibility with existing tidy methods
}
\examples{
gr <- GRanges("chr1:1-10")
gr$p_value <- runif(1)
as_tibble(mcols(gr))
as_tibble(gr)

}
