% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/chopRanges.R
\name{chopRanges}
\alias{chopRanges}
\alias{chopRanges,GRanges-method}
\alias{chopRanges,ANY-method}
\title{Keep unique ranges}
\usage{
chopRanges(x, simplify = TRUE, ...)

\S4method{chopRanges}{GRanges}(x, simplify = TRUE, ...)

\S4method{chopRanges}{ANY}(x, simplify = TRUE, ...)
}
\arguments{
\item{x}{A GenomicRanges object}

\item{simplify}{logical(1). Attempt to simplify returned columns where possible}

\item{...}{Not used}
}
\value{
A GRanges object
}
\description{
Keep unique ranges by 'chopping' mcols
}
\details{
This function finds unique ranges and chops all mcols in a manner similar
to \link[tidyr]{chop}.
Chopped columns will be returned as \code{CompressedList} columns, unless
\code{simplify = TRUE} (the default).
In this case, columns will be returned as vectors where possible.
}
\examples{
gr <- GRanges(rep(c("chr1:1-10"), 2))
gr$id <- paste0("range", seq_along(gr))
gr$gene <- "gene1"
gr
chopRanges(gr)


}
