% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colAsRanges.R
\name{colAsRanges}
\alias{colAsRanges}
\alias{colAsRanges,DataFrame,character-method}
\alias{colAsRanges,GRanges,character-method}
\alias{colAsRanges,data.frame,character-method}
\title{Coerce a column to a GRanges object}
\usage{
colAsRanges(x, var, ...)

\S4method{colAsRanges}{DataFrame,character}(x, var, ...)

\S4method{colAsRanges}{GRanges,character}(x, var, ...)

\S4method{colAsRanges}{data.frame,character}(x, var, ...)
}
\arguments{
\item{x}{A data-frame or GRanges object containing the column to coerce}

\item{var}{The name of the column to coerce}

\item{...}{Not used}
}
\value{
A GenomicRanges object
}
\description{
Coerce a column to a GRanges object
}
\details{
Take a data.frame-like object and coerce one column to a GRanges object,
setting the remainder as the \code{mcols}
}
\examples{
x <- GRanges(c("chr1:1-10", "chr1:6-15", "chr1:51-60"))
df <- data.frame(logFC = rnorm(3), logCPM = rnorm(3,8), p = 10^-rexp(3))
gr <- mergeByCol(x, df, col = "logCPM", pval = "p")
colAsRanges(gr, "keyval_range")

}
