% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colToRanges.R
\name{colToRanges}
\alias{colToRanges}
\alias{colToRanges,DataFrame,character-method}
\alias{colToRanges,GRanges,character-method}
\alias{colToRanges,data.frame,character-method}
\title{Coerce a column to a GRanges object}
\usage{
colToRanges(x, var, ...)

\S4method{colToRanges}{DataFrame,character}(x, var, seqinfo = NULL, ...)

\S4method{colToRanges}{GRanges,character}(x, var, ...)

\S4method{colToRanges}{data.frame,character}(x, var, seqinfo = NULL, ...)
}
\arguments{
\item{x}{A data-frame or GRanges object containing the column to coerce}

\item{var}{The name of the column to coerce}

\item{...}{Not used}

\item{seqinfo}{A seginfo object to be applied to the GRanges object}
}
\value{
A GenomicRanges object
}
\description{
Coerce a column to a GRanges object
}
\details{
Take a data.frame-like object and coerce one column to a GRanges object,
setting the remainder as the \code{mcols}.
A particularly useful application of this is when you have a GRanges object
with one mcol being a secondary GRanges object.

Alternatively, if you have a data.frame with GRanges represented as a
character column, this provides a simple method of coercion.
In this case, no Seqinfo element will be applied to the GRanges element.
}
\examples{
x <- GRanges(c("chr1:1-10", "chr1:6-15", "chr1:51-60"))
df <- data.frame(logFC = rnorm(3), logCPM = rnorm(3,8), p = 10^-rexp(3))
gr <- mergeByCol(x, df, col = "logCPM", pval = "p")
colToRanges(gr, "keyval_range")

}
