% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeByHMP.R
\name{.ec_HMP_adj}
\alias{.ec_HMP_adj}
\title{Similar to the above, this produces the FWER-controlled version in a
streamlined way}
\usage{
.ec_HMP_adj(p, w, L)
}
\arguments{
\item{p}{vector of p-values}

\item{w}{vector of weights}

\item{L}{Number of global tests}
}
\description{
Similar to the above, this produces the FWER-controlled version in a
streamlined way
}
\keyword{internal}
