% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapByFeature.R
\name{.mapFeatures}
\alias{.mapFeatures}
\title{Map ranges to genes using features as an anchor}
\usage{
.mapFeatures(.gr, .feat, .genes, .cols, .gr2feat, .feat2gene, ...)
}
\arguments{
\item{.gr}{The ranges to map onto}

\item{.feat}{Features to use for mapping}

\item{.genes}{GRanges object containing gene-level information}

\item{.cols}{The columns from \code{.genes} to map onto \code{.gr}}

\item{.gr2feat}{The maximum distance between ranges and features}

\item{.feat2gene}{The maximum distance between features & genes}

\item{...}{Passed to findOverlaps and subsetByOverlaps}
}
\value{
A data.frame
}
\description{
Map ranges to genes using features as an anchor
}
\keyword{internal}
