% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapByFeature.R
\name{.mapGi}
\alias{.mapGi}
\title{Map ranges to genes via Interactions}
\usage{
.mapGi(.gr, .gi, .genes, .cols, .gr2gi, .gi2gene, ...)
}
\arguments{
\item{.gr}{The ranges to map onto}

\item{.gi}{GInteractions object}

\item{.genes}{GRanges object containing gene-level information}

\item{.cols}{The columns from \code{.genes} to map onto \code{.gr}}

\item{.gr2gi}{The maximum distance between ranges and anchors}

\item{.gi2gene}{The maximum distance between anchors & genes}

\item{...}{Passed to findOverlaps}
}
\value{
data.frame of mapped ranges
}
\description{
Map ranges to genes via Interactions
}
