% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapByFeature.R
\name{.mapWithin}
\alias{.mapWithin}
\title{Map ranges to all genes within a set distance}
\usage{
.mapWithin(.gr, .genes, .cols, .within, ...)
}
\arguments{
\item{.gr}{The ranges to map onto}

\item{.genes}{GRanges object containing gene-level information}

\item{.cols}{The columns from \code{.genes} to map onto \code{.gr}}

\item{.within}{The maximum distance between ranges & genes}

\item{...}{Passed to findOverlaps}
}
\value{
A data.frame
}
\description{
Map ranges to all genes within a set distance
}
\keyword{internal}
