% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dualFilter.R
\name{dualFilter}
\alias{dualFilter}
\alias{dualFilter,RangedSummarizedExperiment,RangedSummarizedExperiment,GRanges-method}
\alias{dualFilter,ANY,ANY,ANY-method}
\title{Apply two filters to sliding windows}
\usage{
dualFilter(x, bg, ref, ...)

\S4method{dualFilter}{RangedSummarizedExperiment,RangedSummarizedExperiment,GRanges}(
  x,
  bg,
  ref,
  q = 0.5,
  logCPM = TRUE,
  keep.totals = FALSE,
  BPPARAM = bpparam()
)

\S4method{dualFilter}{ANY,ANY,ANY}(x, bg, ref, ...)
}
\arguments{
\item{x}{RangedSummarizedExperiment containing sample counts}

\item{bg}{RangedSummarizedExperiment containing background/input counts}

\item{ref}{GRanges object containing ranges where signal is expected}

\item{...}{Not used}

\item{q}{The upper percentile of the reference ranges expected to be returned
when tuning the filtering criteria}

\item{logCPM}{logical(1) Add a logCPM assay to the returned data}

\item{keep.totals}{logical(1) Keep the original library sizes or replace
using only the retained windows}

\item{BPPARAM}{Settings for running in parallel}
}
\value{
A \link[SummarizedExperiment]{RangedSummarizedExperiment} which is a
filtered subset of the original object. If requested the assay "logCPM" will
be added (\code{TRUE} by default)
}
\description{
Apply two filters to counts generated using sliding windows
}
\details{
This function will take sliding (or tiling) windows for it's input as a
\link[SummarizedExperiment]{RangedSummarizedExperiment} object. The dual
strategy of applying \link[csaw]{filterWindowsControl} and
\link[csaw]{filterWindowsProportion} will then be applied. A set of
reference ranges for which signal is expected is used to refine the
filtering criteria.

Cutoff values are found for both signal relative to input and overall signal,
such that the \verb{100*q\%} of the (sliding) windows which overlap a reference
range will be returned, along with any others which match the
dual filtering criteria.
In general, higher values of \code{q} will return more windows as those with weak
signal and a marginal overlap with a reference range will be returned.
Lower values will ensure that fewer windows, generally with the strongest
signal, are retained.
Cutoff values for both criteria are added to the metadata
element of the returned object.
}
