% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_genes}
\alias{ex_genes}
\title{Basic GRanges containing four genes}
\format{
Simple GRanges object to demonstrate mapping from ranges to genes

All annotations are from GRCh37
}
\usage{
ex_genes
}
\description{
Genes in an example region
}
\examples{
data(ex_genes)
ex_genes
}
\keyword{datasets}
