% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_hic}
\alias{ex_hic}
\title{Basic GInteractions object}
\format{
Simple GInteractions object to demonstrate using features in various
steps implemented in this package

All annotations are from GRCh37
}
\usage{
ex_hic
}
\description{
Contains a single interaction fo demonstration purposes
}
\examples{
data(ex_hic)
ex_hic
}
\keyword{datasets}
