% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_prom}
\alias{ex_prom}
\title{Basic GRanges containing promoters}
\format{
Simple GRanges object to demonstrate using features in various steps
implemented in this package

All annotations are from GRCh37
}
\usage{
ex_prom
}
\description{
Promoters from an example region
}
\examples{
data(ex_prom)
ex_prom
}
\keyword{datasets}
