% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fixed_width_datasets}
\alias{fixed_width_datasets}
\alias{se}
\alias{peaks}
\title{Datasets for the Fixed-Width Vignette}
\format{
An object of class \code{RangedSummarizedExperiment} with 188 rows and 6 columns.

An object of class \code{CompressedGRangesList} of length 6.
}
\usage{
data(se)

data(peaks)
}
\description{
GRangesList of peaks and SummarizedExperiment of counts
All were saved during initial vignette preparation at
https://github.com/smped/extraChIPs_vignette/blob/main/differential_signal_fixed.Rmd
}
\examples{
data(se)
se
data(peaks)
peaks
}
\keyword{datasets}
