% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch37.cytobands}
\alias{grch37.cytobands}
\title{Cytobands for GRCh37/hg19}
\format{
Cytogenetic bands for standard chromosomes from GRCh37,in the format
required by \link[Gviz]{IdeogramTrack}. A data.frame with 862 rows and
5 columns:
\describe{
\item{chrom}{Chromosome}
\item{chromStart}{Starting position for each cytogenetic band}
\item{chromEnd}{End position for each cytogenetic band}
\item{name}{Name for each band, e.g. p.36.33}
\item{gieStain}{Staining pattern}
}
}
\source{
\url{https://hgdownload.soe.ucsc.edu/goldenPath/hg19/database/cytoBand.txt.gz}
}
\usage{
grch37.cytobands
}
\description{
Cytogenetic bands from GRCh37/hg19
}
\examples{
data(grch37.cytobands)
head(grch37.cytobands)

}
\keyword{datasets}
