% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch38.cytobands}
\alias{grch38.cytobands}
\title{Cytobands for GRCh38/hg38}
\format{
Cytogenetic bands for standard chromosomes from GRCh38,in the format
required by \link[Gviz]{IdeogramTrack}. A data.frame with 862 rows and
5 columns:
\describe{
\item{chrom}{Chromosome}
\item{chromStart}{Starting position for each cytogenetic band}
\item{chromEnd}{End position for each cytogenetic band}
\item{name}{Name for each band, e.g. p.36.33}
\item{gieStain}{Staining pattern}
}
}
\source{
\url{https://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/cytoBand.txt.gz}
}
\usage{
grch38.cytobands
}
\description{
Cytogenetic bands from GRCh38/hg38
}
\examples{
data(grch38.cytobands)
head(grch38.cytobands)

}
\keyword{datasets}
