% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConsensus.R
\name{makeConsensus}
\alias{makeConsensus}
\title{Make a set of consensus peaks}
\usage{
makeConsensus(
  x,
  p = 0,
  var = NULL,
  ignore.strand = TRUE,
  simplify = FALSE,
  ...
)
}
\arguments{
\item{x}{A GRangesList}

\item{p}{The minimum proportion of samples (i.e. elements of \code{x}) required
for a peak to be retained in the output. By default all merged peaks will
be returned}

\item{var}{Additional columns in the \code{mcols} element to retain}

\item{ignore.strand, simplify, ...}{Passed to \code{reduceMC()}}
}
\value{
\code{GRanges} object with mcols containing a logical vector for every element of
x, along with the column \code{n} which adds all logical columns.

If any additional columns have been requested using \code{var}, these will be
returned as CompressedList objects as produced by \code{reduceMC()}.
}
\description{
Make a set of consensus peaks based on umber of replicates
}
\details{
This takes a list of GRanges objects and forms a set of consensus peaks
using the minimum proportion of replicates specified
}
\examples{
a <- GRanges("chr1:11-20")
a$score <- 1
b <- GRanges(c("chr1:18-22", "chr1:1-5"))
b$score <- c(0.6, 0.3)
grl <- GRangesList(a = a, b = b)
makeConsensus(grl)
makeConsensus(grl, p = 1)
makeConsensus(grl, p = 1, var = "score")

}
\seealso{
\link{reduceMC}
}
