% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapByFeature.R
\name{mapByFeature}
\alias{mapByFeature}
\title{Map Genomic Ranges to Genes}
\usage{
mapByFeature(
  gr,
  genes,
  prom,
  enh,
  gi,
  cols = c("gene_id", "gene_name", "symbol"),
  range2prom = 0,
  range2enh = 0,
  range2gi = 0,
  range2gene = 1e+05,
  prom2gene = 0,
  enh2gene = 1e+05,
  gi2gene = 0,
  ...
)
}
\arguments{
\item{gr}{GRanges object with ranges to be mapped to genes}

\item{genes}{GRanges object containing genes (or any other nominal feature)
to be assigned}

\item{prom}{GRanges object defining promoters}

\item{enh}{GRanges object defining Enhancers}

\item{gi}{GInteractions object defining interactions. Mappings from
interactions to genes should be performed as a separate prior step.}

\item{cols}{Column names to be assigned as mcols in the output. Columns
must be present in both \code{gi} and \code{genes} if using both objects as
references. Any columns not found in reference objects will be ignored}

\item{range2prom}{The maximum permissible distance between a query range and
any ranges defined as promoters}

\item{range2enh}{The maximum permissible distance between a query range and
any ranges defined as enhancers}

\item{range2gi}{The maximum permissible distance between a query range and
any ranges defined as GInteraction anchors}

\item{range2gene}{The maximum permissible distance between a query range and
genes (for ranges not otherwise mapped)}

\item{prom2gene}{The maximum permissible distance between a range provided
in \code{prom} and a gene}

\item{enh2gene}{The maximum permissible distance between a range provided
in \code{enh} and a gene}

\item{gi2gene}{The maximum permissible distance between a GInteractions
anchor (provided in \code{gi}) and a gene}

\item{...}{Passed to findOverlaps nad overlapsAny internally}
}
\value{
A GRanges object with added mcols as specified
}
\description{
Map Genomic Ranges to Genes using Key Features
}
\details{
This function is able to incorporate feature-level information and long-range
interactions to enable better mapping of regions to genes. For GRanges, the
following sequential strategy is used.

\enumerate{
\item Ranges overlapping a promoter are assigned to that gene
\item Ranges overlapping an enhancer are assigned to all genes within
a specified distance
\item Ranges overlapping a long-range interaction are assigned to all genes
connected by the interaction
\item Ranges with no gene assignment from the previous steps are assigned
to the nearest gene within a specified distance
}

For long-range interactions, the above strategy is again used, taking each
anchor as a distinct range. Mappings for both anchors are then combined for
interaction-specific mappings
}
