% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeByCol.R
\name{mergeByCol}
\alias{mergeByCol}
\alias{mergeByCol,GenomicRanges-method}
\alias{mergeByCol,RangedSummarizedExperiment-method}
\title{Merge Sliding Windows}
\usage{
mergeByCol(x, ...)

\S4method{mergeByCol}{GenomicRanges}(
  x,
  df = NULL,
  col,
  by = c("max", "median", "mean", "min"),
  logfc = "logFC",
  pval = "P",
  inc_cols,
  p_adj_method = "fdr",
  merge_within = 1L,
  ignore_strand = FALSE,
  ...
)

\S4method{mergeByCol}{RangedSummarizedExperiment}(
  x,
  df = NULL,
  col,
  by = c("max", "median", "mean", "min"),
  logfc = "logFC",
  pval = "P",
  inc_cols,
  p_adj_method = "fdr",
  merge_within = 1L,
  ignore_strand = FALSE,
  ...
)
}
\arguments{
\item{x}{A GenomicRanges or SummarizedExperiment object}

\item{...}{Not used}

\item{df}{A data.frame-like object containing the columns of interest. If
not provided, any columns in the mcols() slot will be used.}

\item{col}{The column to select as representative of the merged ranges}

\item{by}{The method for selecting representative values}

\item{logfc}{Column containing logFC values}

\item{pval}{Column containing p-values}

\item{inc_cols}{Any additional columns to return. Output will always include
columns specified in the arguments \code{col}, \code{logfc} and \code{pval}}

\item{p_adj_method}{Any of \link{p.adjust.methods}}

\item{merge_within}{Merge any ranges within this distance}

\item{ignore_strand}{Passed internally to \link[GenomicRanges]{reduce} and
\link[GenomicRanges]{findOverlaps}}
}
\value{
A Genomic Ranges object
}
\description{
Merge Sliding Windows by a given column
}
\details{
This merges sliding windows by a given column, returning representative
values from each requested column, corresponding to the selected row.
Values from the requested column can be selected using any of \code{min()},
\code{max()}, \code{mean()} or \code{median()}, although \code{max()} is strongly recommended.
The range corresponding to the returned values is returned as the column.
\code{keyval_range}.

Taking the p-value from the selected row, the number of windows with lower
p-values are counted by direction and returned in the final object
}
\examples{
x <- GRanges(c("chr1:1-10", "chr1:6-15", "chr1:51-60"))
df <- data.frame(logFC = rnorm(3), logCPM = rnorm(3,8), p = 10^-rexp(3))
mergeByCol(x, df, col = "logCPM", pval = "p")

}
