% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapsProp.R
\name{overlapsProp,GRanges,GRanges-method}
\alias{overlapsProp,GRanges,GRanges-method}
\alias{overlapsProp}
\alias{overlapsProp,ANY,ANY-method}
\title{Find the proportions of an overlapping range}
\usage{
\S4method{overlapsProp}{GRanges,GRanges}(x, y, ignore.strand = FALSE, ...)

\S4method{overlapsProp}{ANY,ANY}(x, y, ignore.strand = FALSE, ...)
}
\arguments{
\item{x, y}{A GenomicRanges object}

\item{ignore.strand}{If set to TRUE, then the strand of x and y is set to
"*" prior to any computation.}

\item{...}{Not used}
}
\value{
Numeric vector the same length as x
}
\description{
Find the proportion of a query reange which overlaps the subject
}
\details{
This behaves similarly to \link[IRanges]{overlapsAny} except the proportion
of the query range which overlaps one or more subject ranges is returned
instead of a logical vector
}
\examples{
x <- GRanges("chr1:1-10")
y <- GRanges("chr1:1-5")
overlapsProp(x, y)
overlapsProp(y, x)

}
