% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPie.R
\name{plotPie}
\alias{plotPie}
\title{Draw Pie Graphs based on one or more columns}
\usage{
plotPie(
  df,
  fill,
  x,
  y,
  width = 0.8,
  show_total = TRUE,
  label_fill = "white",
  label_alpha = 1,
  label_size = 3,
  min_p = 0.01,
  show_category = TRUE,
  category_size = 3,
  category_colour = "black",
  category_width = 15,
  ...
)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{fill}{The category/column used to fill the slices of the pie charts}

\item{x}{The second (optional) category/column to place along the x-axis}

\item{y}{The final (optional) category/column to plce along the y-axis}

\item{width}{Scale the width of all charts}

\item{show_total}{logical(1) Show labels on each pie chart with the tally for
that complete chart}

\item{label_fill}{The backgrund colour for tally labels}

\item{label_alpha}{Transparency for taly labels}

\item{label_size}{Size of the tally labels. Passed to
\link[ggplot2]{geom_label}}

\item{min_p}{The minimum proportion of the total required for adding labels.
Effectively removes labels from pie charts with few members. Alternatively
when only one column is specified, categories below this will not be shown
around the adge of the plot}

\item{show_category}{Show category labels around the edge of the plot if only
one category/column is specified}

\item{category_size}{The size of category labels if only one category/column
is specified}

\item{category_colour}{The colour of category labels if only one column is
specified}

\item{category_width}{Width at which category labels will wrap onto a new
line}

\item{...}{Not used}
}
\value{
A ggplot2 object able to be customised with colour scales and themes
}
\description{
Draw Pie Graphs based one or more data.frame columns
}
\details{
Using a \code{data.frame} as input, this function will draw pie graphs based
on one ore more columns, by simply counting the values in combination
across these columns.
One column must be selected for the fill as a bare minimum, with up to three
being possible.
Additional columns can be set for the x-axis to draw a series of pie-graphs
in a row, with a further column able to added to layout a series of pie
graphs in a grid

If only one column/category is chosen, category labels will be added around
the edge of the plot

If \code{show_total = TRUE} the overall counts for each pie graph will be added
in the centre using \link[ggplot2]{geom_label}.
Parameters for these labels are customisable
}
\examples{
set.seed(200)
df <- data.frame(
  feature = sample(c("Promoter", "Enhancer", "Intergenic"), 200, replace = TRUE),
  TF1 = sample(c("Up", "Down", "Unchanged"), 200, replace = TRUE),
  TF2 = sample(c("Up", "Down", "Unchanged"), 200, replace = TRUE)
)
plotPie(df, fill = "feature")
plotPie(df, fill = "feature", x = "TF1")
plotPie(df, fill = "feature", x = "TF1", y = "TF2") +
 scale_fill_viridis_d() +
 theme_bw()

}
