% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfileHeatmap.R
\name{plotProfileHeatmap}
\alias{plotProfileHeatmap}
\title{Draw a Coverage Profile Heatmap}
\usage{
plotProfileHeatmap(
  profiles,
  ids,
  valueCol = "score",
  valueLab = "Relative\\nCoverage",
  xCol = "bp",
  xLab = "Distance from Centre (bp)",
  addHist = TRUE,
  relHeights = c(0.3, 0.7),
  scaleTo = c("median", "mean", "max", "min", "none"),
  ...
)
}
\arguments{
\item{profiles}{A set of profile data.frames, or a list of these data frames.
See details.}

\item{ids}{The IDs for the profiles. Most commonly the range as a character
vector}

\item{valueCol}{The column in each profile data.frame with the coverage value}

\item{valueLab}{The label to use for the legend, e.g. CPM, Coverage etc}

\item{xCol}{The column name to be used for the x-axis}

\item{xLab}{The label for the x-axis}

\item{addHist}{logical(1). Add a line histogram to the top of the heatmap}

\item{relHeights}{The relative heights of the histogram and heatmap. Ignored
if addHist = \code{FALSE}}

\item{scaleTo}{Scale values to the median, mean etc of the provided profile
data frame. Currently, this value will be subtracted from the score for
greater comparability across multiple samples within the same heatmap}

\item{...}{Not used}
}
\value{
A \code{ggplot2} object, able to be customised using standard \code{ggplot2} syntax
}
\description{
Plot a Coverage Profile Heatmap across multiple ranges
}
\details{
Convenience function for plotting coverage heatmaps across a common set of
ranges, shared between one or more samples. These are most commonly the
coverage values from merged samples within a treatment group.

A profile data frame here refers to a data.frame (or tibble, or DataFrame)
with a coverage value in one column that corresponds to a genomic bin of a
fixed size, as generated by \link{getProfileData}. Given that multiple
ranges are most likely to be drawn, each profile data frame must be the same
size in terms of the number of bins, each of which represent a fixed number
of nucleotides.

If providing a list of profile data frames, the plot will drawn with each
list element as a separate panel within the larger heatmap, which assumes
each element to represent a sample group. The order these are passed to the
function will be the order in which they are drawn

By default (addHist = \code{TRUE}) a summarised histogram will be drawn at the
top of each sample by taking the mean coverage within each genomic bin.
}
\examples{
bw <- system.file("tests", "test.bw", package = "rtracklayer")
gr <- GRanges("chr2:500")
pd <- getProfileData(bw, gr, upstream = 100, bins = 10)
## This looks terrible given it's a single range with even coverage
## but the syntax should be helpful
plotProfileHeatmap(pd$profile_data, ids = granges(pd))

}
