% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangesAsTibble.R
\name{rangesAsTibble}
\alias{rangesAsTibble}
\alias{mcolsAsTibble}
\title{Convert GRanges to a tibble}
\usage{
rangesAsTibble(x, var = "range", drop.existing = FALSE)

mcolsAsTibble(x)
}
\arguments{
\item{x}{A Genomic Ranges object}

\item{var}{Name of column to use for ranges}

\item{drop.existing}{logical(1). Overwrite any existing columns with the
same name as supplied in var}
}
\value{
A \link[tibble]{tibble}
}
\description{
Convert GRanges objects to tibble objects
}
\details{
Simple conversion of a GRanges object to a tibble. The ranges will be
returned as a character column called \code{ranges}. Seqinfo information will be
lost.
}
\examples{
gr <- GRanges("chr1:1-10")
gr$p_value <- runif(1)
rangesAsTibble(gr)
mcolsAsTibble(gr)

}
