% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangesToGenes.R
\name{rangesToGenes}
\alias{rangesToGenes}
\title{Map Genomic Ranges to Genes}
\usage{
rangesToGenes(
  gr,
  prom,
  enh,
  hic,
  genes,
  cols = c("gene_id", "gene_name", "symbol"),
  maxgap = 0,
  maxdist = 1e+05,
  ...
)
}
\arguments{
\item{gr}{GRanges object with ranges to be mapped to genes}

\item{prom}{GRanges object defining promoters}

\item{enh}{GRanges object defining Enhancers}

\item{hic}{GInteractions object defining interactions. Mappings from
interactions to genes should be performed as a separate prior step.}

\item{genes}{GRanges object containing genes (or any other nominal feature)
to be assigned}

\item{cols}{Column names to be assigned as mcols in the output. Columns
must be present in both \code{hic} and \code{genes} if using both objects as
references. Any columns not found in reference objects will be ignored}

\item{maxgap}{The maximum distance between range and features to be
considered as a direct overlap}

\item{maxdist}{The maximum distance between ranges/enhancers and genes for
mapping any otherwise unmapped ranges}

\item{...}{Not used}
}
\value{
A GRanges object with added mcols as specified
}
\description{
Map Genomic Ranges to Genes using Multiple Data Sources
}
\details{
This function is able to incorporate feature-level information and HiC
interactions to enable better mapping of regions to genes. For GRanges, the
following sequential strategy is used.

\enumerate{
\item Ranges overlapping a promoter are assigned to that gene
\item Ranges overlapping an enhancer are assigned to all genes within
a specified distance
\item Ranges overlapping a HiC interaction are assigned to all genes
connected by the interaction
\item Ranges with no gene assignment from the previous steps are assigned
to the nearest gene within a specified distance
}

For HiC Interactions, the above strategy is again used, excluding the HiC
step and taking each anchor as a distinct range. Mappings for both anchors
are then combined for interaction-specific mappings
}
