% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiffWithMCols.R
\name{setdiffWithMcols}
\alias{setdiffWithMcols}
\title{Perform setdiff operations retaining mcols}
\usage{
setdiffWithMcols(x, y, ignore.strand = FALSE, ...)
}
\arguments{
\item{x, y}{GenomicRanges objects}

\item{ignore.strand}{If set to TRUE, then the strand of x and y is set to
"*" prior to any computation.}

\item{...}{Not used}
}
\value{
A GRanges object with all mcols returned form the original object.
If a range obtained by setdiff maps back to two or more ranges in the
original set of Ranges, mcols will be returned as
\link[IRanges]{CompressedList} or
\link[S4Vectors]{SimpleList} columns
}
\description{
Perform setdiff operations retaining all mcols from the query range
}
\details{
This extends the methods provided by \link[GenomicRanges]{setdiff}
}
\examples{
x <- GRanges("chr1:1-100:+")
x$id <- "range1"
y <- GRanges(c("chr1:51-60:+", "chr1:21-30:-"))
setdiffWithMcols(x, y)
setdiffWithMcols(x, y, ignore.strand = TRUE)

}
