% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setoptsMC.R
\name{setdiffMC,GRanges,GRanges-method}
\alias{setdiffMC,GRanges,GRanges-method}
\alias{setdiffMC}
\alias{setdiffMC,ANY,ANY-method}
\title{Perform set operations retaining mcols}
\usage{
\S4method{setdiffMC}{GRanges,GRanges}(x, y, ignore.strand = FALSE, simplify = TRUE, ...)

\S4method{setdiffMC}{ANY,ANY}(x, y, ignore.strand = FALSE, simplify = TRUE, ...)
}
\arguments{
\item{x, y}{GenomicRanges objects}

\item{ignore.strand}{If set to TRUE, then the strand of x and y is set to
"*" prior to any computation.}

\item{simplify}{logical(1) If TRUE, any List columns will be returned as
vectors where possible. This can only occur if single, unique entries are
present in all initial elements.}

\item{...}{Not used}
}
\value{
A GRanges object with all mcols returned form the original object.
If a range obtained by setdiff maps back to two or more ranges in the
original set of Ranges, mcols will be returned as
\link[IRanges]{CompressedList} or
\link[S4Vectors]{SimpleList} columns
}
\description{
Perform set operations retaining all mcols from the query range
}
\details{
This extends the methods provided by \link[GenomicRanges]{setdiff}
}
\examples{
x <- GRanges("chr1:1-100:+")
x$id <- "range1"
y <- GRanges(c("chr1:51-60:+", "chr1:21-30:-"))
setdiffMC(x, y)
setdiffMC(x, y, ignore.strand = TRUE)

}
