% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQSmooth.R
\name{addQSmooth}
\alias{addQSmooth}
\alias{addQSmooth,SummarizedExperiment-method}
\title{Smooth Quantile Normalise a SummarizedExperiment}
\usage{
addQSmooth(x, ...)

\S4method{addQSmooth}{SummarizedExperiment}(x, assay = "counts", factor, ...)
}
\arguments{
\item{x}{a RangedSummarizedExperiment object}

\item{...}{Passed to \link[qsmooth]{qsmooth}}

\item{assay}{The assay to apply SmoothQuantileNormalisation to}

\item{factor}{A factor used to group the data. See \link[qsmooth]{qsmooth}}
}
\value{
A SummarizedExperiment with additional metadata and the qsmooth assay
}
\description{
Smooth Quantile Normalise a SummarizedExperiment
}
\details{
This function adds the assay 'qsmooth' to a SummarizedExperiment or any
derivative class, choosing which existing assay to normalise.

The qsmooth weights are returned in the metadata element of the final object.
}
