% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineRegions.R
\name{defineRegions}
\alias{defineRegions}
\title{Define Genomic Regions Based on Gene Annotations}
\usage{
defineRegions(
  genes,
  transcripts,
  exons,
  promoter = c(2500, 500),
  upstream = 5000,
  intron = TRUE,
  proximal = 10000,
  cols = c("gene_id", "gene_name", "transcript_id", "transcript_name"),
  ...
)
}
\arguments{
\item{genes, transcripts, exons}{GRanges objects defining each level of
annotation}

\item{promoter}{Numeric vector defining upstream and/or downstream distances
for a promoter. Passing a single value will define a symmetrical promoter
The first value represents the upstream range}

\item{upstream}{The distance from a TSS defining an upstream promoter}

\item{intron}{logical(1) Separate gene bodies into introns and exons. If
\code{intron = FALSE} gene bodies will simply be defined as gene bodies}

\item{proximal}{Distance from a gene to be considered a proximal intergenic
region. If set to 0, intergenic regions will simply be considered as
uniformly intergenic}

\item{cols}{Column names to be retained from the supplied annotations}

\item{...}{Not used}
}
\value{
A GRangesList
}
\description{
Use gene, transcript and exon annotations to define genomic regions
}
\details{
Using GRanges annotated as genes, transcripts and exons this function will
define ranges uniquely assigned to a region type using a hierarchical
process. By default, these region types will be (in order) 1) Promoters, 2)
Upstream Promoters, 3) Exons, 4) Introns, 5) Proximal Intergenic and 6)
Distal Intergenic.

Setting \code{intron = FALSE} will replace introns and exons with a generic "Gene
Body" annotation.
Setting \code{proximal = 0} will return all intergenic regions (not previously
annotated as promoters or upstream promoters) to an "Intergenic" category

Notably, once a region has been defined, it is excluded from all subsquent
candidate regions.

Any columns matching the names provided in cols will be returned, and it is
assumed that the gene/transcript/exon ranges will contain informative columns
in the \code{mcols()} element.
}
