% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/distincMC.R
\name{distinctMC}
\alias{distinctMC}
\alias{distinctMC,GRanges-method}
\alias{distinctMC,ANY-method}
\title{Keep distinct ranges and mcols}
\usage{
distinctMC(x, .across = everything(), ...)

\S4method{distinctMC}{GRanges}(x, .across = everything(), ...)

\S4method{distinctMC}{ANY}(x, .across = everything(), ...)
}
\arguments{
\item{x}{A GenomicRanges object}

\item{.across}{\verb{<tidy-select>} Passed to \link[dplyr]{across}}

\item{...}{Not used}
}
\value{
A GRanges object
}
\description{
Keep distinct ranges by including mcols
}
\details{
This function finds unique ranges and mcols in combination and retains only
the distinct combinations
}
\examples{
library(tidyselect)
gr <- GRanges(rep(c("chr1:1-10"), 2))
gr$id <- paste0("range", seq_along(gr))
gr$gene <- "gene1"
gr
distinctMC(gr)
distinctMC(gr, all_of("gene"))

}
