% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/distinctMC.R
\name{distinctMC}
\alias{distinctMC}
\alias{distinctMC,GRanges-method}
\title{Keep distinct ranges and mcols}
\usage{
distinctMC(x, cols, ...)

\S4method{distinctMC}{GRanges}(x, cols, ...)
}
\arguments{
\item{x}{A GenomicRanges object}

\item{cols}{The mcols to be used to determining distinct ranges/mcols}

\item{...}{Not used}
}
\value{
A GRanges object
}
\description{
Keep distinct ranges by including mcols
}
\details{
This function finds unique ranges and mcols in combination and retains only
the distinct combinations. If no mcols are present defaults to \code{unique(x)}
}
\examples{
library(tidyselect)
gr <- GRanges(rep(c("chr1:1-10"), 2))
gr$id <- paste0("range", seq_along(gr))
gr$gene <- "gene1"
gr
distinctMC(gr)
distinctMC(gr, "gene")

}
