% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualFilter.R
\name{dualFilter}
\alias{dualFilter}
\title{Apply two filters to sliding windows}
\usage{
dualFilter(x, bg, ref, q = 0.99, logCPM = TRUE, BPPARAM = bpparam())
}
\arguments{
\item{x}{RangedSummarizedExperiment containing sample counts}

\item{bg}{RangedSummarizedExperiment containing background/input counts}

\item{ref}{GRanges object containing ranges where signal is expected}

\item{q}{The upper percentile of the reference ranges expected to be returned
when tuning the filtering criteria}

\item{logCPM}{logical(1) Add a logCPM assay to the returned data}

\item{BPPARAM}{Settings for running in parallel}
}
\value{
A \link[SummarizedExperiment]{RangedSummarizedExperiment} which is a
filtered subset of the original object. If requested the assay "logCPM" will
be added
}
\description{
Apply two filters to counts generated using sliding windows
}
\details{
This function will take sliding (or tiling) windows for it's input as a
\link[SummarizedExperiment]{RangedSummarizedExperiment} object. The dual
strategy of applying \link[csaw]{filterWindowsControl} and
\link[csaw]{filterWindowsProportion} will then be applied. A set of
reference ranges for which signal is expected is used to refine the
filtering criteria.

Cutoff values are found such that the \code{q} of the windows which overlap one of
the reference ranges will be returned, along with any others which match the
filtering criteria
}
