% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPeaks.R
\name{importPeaks}
\alias{importPeaks}
\title{Import peaks}
\usage{
importPeaks(
  x,
  type = c("narrow", "broad", "bed"),
  blacklist,
  seqinfo,
  pruning.mode = c("coarse", "error"),
  sort = TRUE,
  setNames = TRUE,
  centre = FALSE,
  nameRanges = TRUE,
  ...
)
}
\arguments{
\item{x}{One or more files to be imported. All files must be of the same
type, i.e. narrow or broad}

\item{type}{The type of peaks to be imported. All files are assumed to be
the same format}

\item{blacklist}{A set of ranges to be excluded}

\item{seqinfo}{A seqinfo object to be applied to the GRanges objects}

\item{pruning.mode}{How to handle conflicts if supplying a seqinfo object.
Defaults to \code{pruning.mode = "coarse"}. Only "coarse" and "error" are
implemented. See \link[GenomeInfoDb]{seqinfo}.}

\item{sort}{logical. Should the ranges be sorted during import}

\item{setNames}{logical Set basename(x) as the name}

\item{centre}{Add the position of the peak centre. Ignored unless
importing narrowPeak files.}

\item{nameRanges}{Set the values in the 'name' column as the range name.
Ignored if parsing a bed file}

\item{...}{passed to \code{sort}}
}
\value{
A GRangesList
}
\description{
Import peaks in narrowPeak or broadPeak format
}
\details{
Peaks are imported from either narrowPeak, broadPeak or bed format as
individual GenomicRanges objects, before being incorporated as a single
GRangesList. If importing bed files, \link[rtracklayer]{import.bed} will be
called internally.
}
\examples{
fl <- system.file(
    c("extdata/ER_1.narrowPeak", "extdata/ER_2.narrowPeak"),
    package = "extraChIPs"
)
peaks <- importPeaks(fl)
peaks

}
