% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPeaks.R
\name{importPeaks}
\alias{importPeaks}
\title{Import peaks}
\usage{
importPeaks(
  x,
  type = c("narrow", "broad"),
  blacklist,
  seqinfo,
  pruning.mode = c("coarse", "error"),
  sort = TRUE,
  ...
)
}
\arguments{
\item{x}{One or more files to be imported. All files must be of the same
type, i.e. narrow or broad}

\item{type}{The type of peaks to be imported}

\item{blacklist}{A set of ranges to be excluded}

\item{seqinfo}{A seqinfo object to be applied to the GRanges objects}

\item{pruning.mode}{How to handle conflicts if supplying a seqinfo object.
Defaults to \code{pruning.mode = "coarse"}. Only "coarse" and "error" are
implemented. See \link[GenomeInfoDb]{seqinfo}.}

\item{sort}{logical. Should the ranges be sorted during import}

\item{...}{passed to \code{sort}}
}
\value{
A GRangesList
}
\description{
Import peaks in narrowPeak or broadPeak format
}
\details{
Peaks are imported from either narrowPeak or broadPeak format as
GenomicRanges objects.
}
\examples{
fl <- system.file(
"extdata", "testFiles", "test.narrowPeak", package = "chipExtra"
)
gr <- importPeaks(fl, "narrow")

}
