% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGrlCols.R
\name{mapGrlCols}
\alias{mapGrlCols}
\title{Collapse a GRangesList adding multiple columns from each element}
\usage{
mapGrlCols(
  x,
  var = NULL,
  collapse = NULL,
  collapse_sep = "; ",
  name_sep = "_",
  ...
)
}
\arguments{
\item{x}{GRangesList}

\item{var}{columns to map onto the set of consensus peaks}

\item{collapse}{Any columns specified here will be simplified into a single
column}

\item{collapse_sep, }{String to separate values whe collapsing columns}

\item{name_sep}{String to separate values when add ing column names}

\item{...}{Passed to makeConsensus}
}
\value{
GRanges object
}
\description{
Make consensus peaks and add individual columns from each original
GRangesList element
}
\details{
Starting with a GRangesList, make a single GRanges object with select columns
from each element added to the new object
}
