% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAssayDensities.R
\name{plotAssayDensities}
\alias{plotAssayDensities}
\alias{plotAssayDensities,SummarizedExperiment-method}
\title{Plot Densities for any assay within a SummarizedExperiment}
\usage{
plotAssayDensities(x, ...)

\S4method{plotAssayDensities}{SummarizedExperiment}(
  x,
  assay = "counts",
  group,
  colour = NULL,
  linetype = NULL,
  trans = NULL,
  ...
)
}
\arguments{
\item{x}{A SummarizedExperiment object}

\item{...}{Passed to \link[stats]{density}}

\item{assay}{An assay within x}

\item{group}{Draw lines based on this column in colData. Defaults to
each sample}

\item{colour}{The column in colData to colour lines by}

\item{linetype}{Any column in colData used to determine linetype}

\item{trans}{character(1). Any transformative function to be applied to the
data before calculating the density, e.g. \code{trans = "log2"}}
}
\value{
A \code{ggplot2} object. Scales and labels can be added using conventional
\code{ggplot2} syntax. (See example)
}
\description{
Plot Densities for any assay within a SummarizedExperiment
}
\details{
Uses ggplot2 to create a density plot for all samples within the selected
assay
}
\examples{
nrows <- 200; ncols <- 4
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
df <- DataFrame(treat = c("A", "A", "B", "B"))
se <- SummarizedExperiment(
  assays = SimpleList(counts = counts),
  colData = df
)
plotAssayDensities(se, colour = "treat") +
  labs(colour = "Treat")

}
