% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAssayDensities.R
\name{plotAssayDensities}
\alias{plotAssayDensities}
\alias{plotAssayDensities,SummarizedExperiment-method}
\title{Plot Densities For Any Assay}
\usage{
plotAssayDensities(x, ...)

\S4method{plotAssayDensities}{SummarizedExperiment}(x, assay = "counts", colour_by = c(), group_by, ...)
}
\arguments{
\item{x}{A SummarizedExperiment object}

\item{...}{Passed to \link[stats]{density}}

\item{assay}{An assay within x}

\item{colour_by}{The column in colData to colour lines by}

\item{group_by}{Draw lines based on this column in colData. Defaults to
each sample}
}
\value{
A ggplot2 object
}
\description{
Plot Densities for any assay within a SummarizedExperiment
}
\details{
Uses ggplot2 to create a density plot for all samples within the selected
assay
}
\examples{
nrows <- 200; ncols <- 4
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
df <- DataFrame(treat = c("A", "A", "B", "B"))
se <- SummarizedExperiment(
  assays = SimpleList(counts = counts),
  colData = df
)
plotAssayDensities(se, colour_by = "treat")

}
