% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotAssayRle.R
\name{plotAssayRle}
\alias{plotAssayRle}
\alias{plotAssayRle,SummarizedExperiment-method}
\title{Plot RLE for a given assay within a SummarizedExperiment}
\usage{
plotAssayRle(x, ...)

\S4method{plotAssayRle}{SummarizedExperiment}(
  x,
  assay = "counts",
  colour = NULL,
  fill = NULL,
  rle_group = NULL,
  n_max = Inf,
  trans = c(),
  ...
)
}
\arguments{
\item{x}{A SummarizedExperiment object}

\item{...}{Not used}

\item{assay}{The assay to plot}

\item{colour}{Column from \code{colData(x)} to outline the boxplots}

\item{fill}{Column from \code{colData(x)} to fill the boxplots}

\item{rle_group}{Column from \code{colData(x)} to calculate RLE within groups
Commonly an alternative sample label.}

\item{n_max}{Maximum number of points to plot}

\item{trans}{character(1). NUmerical transformation to apply to the data
prior to RLE calculation}
}
\value{
A ggplot2 object
}
\description{
Plot RLE for a given assay within a SummarizedExperiment
}
\details{
Uses ggplot2 to create an RLE plot for the selected assay. Any numerical
transformation prior to performing the RLE can be specified using the
\code{trans} argument
}
\examples{
nrows <- 200; ncols <- 4
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
df <- DataFrame(treat = c("A", "A", "B", "B"))
se <- SummarizedExperiment(
  assays = SimpleList(counts = counts),
  colData = df
)
plotAssayRle(se, "counts", fill = "treat")

}
