% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHFGC.R
\name{plotHFGC}
\alias{plotHFGC}
\title{Plot a Genomic Region}
\usage{
plotHFGC(
  gr,
  hic,
  features,
  genes,
  coverage,
  zoom = 1,
  shift = 0,
  max = 1e+07,
  axistrack = TRUE,
  cytobands,
  coverage_type = c("l", "heatmap"),
  linecol,
  gradient = hcl.colors(101, "viridis"),
  hiccol = list(anchors = "lightblue", interactions = "red"),
  featurecol,
  genecol,
  highlight = "blue",
  hicsize = 1,
  featsize = 1,
  genesize = 1,
  covsize = 4,
  ...,
  fontsize = 12,
  cex.title = 0.8,
  rotation.title = 0,
  collapseTranscripts = "meta"
)
}
\arguments{
\item{gr}{The range(s) of interest. Must be on a single chromosome}

\item{hic}{Any HiC interactions to be included as a GenomicInteractions
object. If not supplied, no HiC track will be drawn.}

\item{features}{A named GRangesList object containing regulatory features in
each list element. Features will be drawn on a single track with colours
matching those provided in \code{featurecol}. If not included, no feature track
will be drawn}

\item{genes}{A GRanges object with exon structure for each transcript/gene.
If not included, to track will be drawn for gene/transcript structure}

\item{coverage}{A named list of BigWigFileList objects containing the
primary tracks to show coverage for. Each list element will be drawn on a
separate track, with elements within each BigWigFileList shown on the same
track. List names will become track names. Alternatively, a single
BigWigFileList will plot all individual files on separate tracks.
If not included, no coverage tracks will be drawn.}

\item{zoom}{Multiplicative factor for zooming in and out}

\item{shift}{Shift the plot. Applied after zooming}

\item{max}{The maximum width of the plotting region. Given that the width of
the final plotting window will be determined by any HiC interactions, this
argument excludes any interactions beyond this distance. Plotting can be
somewhat slow if any long range interactions are included. Ignored if no HiC
interactions are supplied.}

\item{axistrack}{logical. Add an AxisTrack()}

\item{cytobands}{Cytogenetic bands to be displayed on each chromosome}

\item{coverage_type}{The plot type for coverage. Currently only lines ("l")
and heatmaps ("heatmap") are supported}

\item{linecol}{Named vector (or list) of colours for each coverage track.
Only used if coverage_type = "l". All named elements within the coverage
object should be included.}

\item{gradient}{Colour gradient for heatmaps}

\item{hiccol}{list with names \code{"anchors"} and \code{"interactions"}. Colours
are passed to these elements}

\item{featurecol}{Named vector (or list) of colours for each feature}

\item{genecol}{Named vector (or list) of colours for each gene category}

\item{highlight}{Outline colour for the highlight track. Setting this to
\code{NULL} will remove the highlight}

\item{hicsize, featsize, genesize, covsize}{Relative sizes for each track}

\item{...}{Passed to \link[Gviz]{DataTrack} for the \strong{coverage tracks} only.
Useful arguments may be things like \code{ylim}, \code{legend}}

\item{fontsize}{Applied across all tracks}

\item{cex.title}{Passed to all tracks}

\item{rotation.title}{Passed to all tracks}

\item{collapseTranscripts}{Passed to \link[Gviz]{GeneRegionTrack} for the
genes track}
}
\value{
A Gviz object
}
\description{
Plot a region with HiC, Features, Genes and Coverage
}
\details{
Convenience function for plotting a common set of tracks. All tracks are
optional. For more fine control, users are advised to simply use Gviz
directly.

The primary tracks defined in this function are H (HiC), F (features), G
(genes), and C (coverage). Axis and Ideogram tracks are an additional part of
this visualisation

Use all tracks specific to this dataset to generate a simple visualisation.
In descending order the tracks displayed will be:
\enumerate{
\item HiC Interactions (if supplied)
\item Regulatory features
\item Genes/genes
\item Coverage tracks as supplied
}

All tracks are optional and will simply be omitted if no data is supplied.

If wanting a single track of genes, simply pass a GRanges object in the
format specified for a \link[Gviz]{GeneRegionTrack}. Passing a GRangesList
with the same format will yield an individual track for each list element,
with each track shown by default as a separate colour. This can be used for
showing Up/Down-regulated genes, or Detected/Undetected genes.

If passing a BigWigFileList for the coverage track, each file within the
object will be drawn on a separate track. If passing a list of BigWigFileList
objects, each list element will be drawn as a single track with the
individual files within each BigWigFileList overlaid within each track.

Cytogenetic band information must be in the structure required by
\link[Gviz]{IdeogramTrack}, with data for both GRCh37 and GRCh38 provided in
this package (\link{grch37.cytobands}, \link{grch38.cytobands}).

A highlight overlay over the GRanges provided as the \code{gr} argument will be
added if a colour is provided. If set to NULL, no highlight will be added.
}
\examples{
gr <- GRanges("chr1:11869-12227")
feat_gr <- GRangesList(
  Promoter = GRanges("chr1:11800-12000"),
  Enhancer = GRanges("chr1:13000-13200")
)
hic <- InteractionSet::GInteractions(feat_gr$Promoter, feat_gr$Enhancer)
genes <- c("chr1:11869-12227:+", "chr1:12613-12721:+", "chr1:13221-14409:+")
genes <- GRanges(genes)
mcols(genes) <- DataFrame(
  feature = "exon", gene = "ENSG00000223972", exon = 1:3,
  transcript = "ENST00000456328", symbol = "DDX11L1"
)
data(grch37.cytobands)
plotHFGC(
  gr, hic = hic, features = feat_gr, genes = genes,
  zoom = 2, cytobands = grch37.cytobands, rotation.title = 90,
  featurecol = c(Promoter = "red", Enhancer = "yellow")
)

}
