% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfileHeatmap.R
\name{plotProfileHeatmap}
\alias{plotProfileHeatmap}
\alias{plotProfileHeatmap,GenomicRanges-method}
\title{Draw a Coverage Profile Heatmap}
\usage{
plotProfileHeatmap(object, ...)

\S4method{plotProfileHeatmap}{GenomicRanges}(
  object,
  profileCol,
  xValue = "bp",
  fillValue = "score",
  facetX = NULL,
  facetY = NULL,
  colour = facetY,
  linetype = NULL,
  summariseBy = c("mean", "median", "min", "max", "none"),
  xLab = xValue,
  yLab = NULL,
  fillLab = fillValue,
  facetLabeller = "label_value",
  relHeight,
  ...
)
}
\arguments{
\item{object}{A GRanges object}

\item{...}{Not used}

\item{profileCol}{Where to find the profile DataFrames}

\item{xValue, fillValue}{Columns within the profile DataFrames for heatmaps}

\item{facetX}{Column used for facetting across the x-axis}

\item{facetY}{Column used for facetting across the y-axis}

\item{colour}{Column used for colouring lines in the summary panel. Defaults
to the column used for facetY}

\item{linetype}{Column used for linetypes in the summary panel}

\item{summariseBy}{FUnction for creating the summary plot in the top panel.
If set to 'none', no summary plot will be drawn}

\item{xLab, yLab, fillLab}{Labels for plotting aesthetics. Can be overwritten
using labs()}

\item{facetLabeller}{Any labeller function able to be passed to facet_grid}

\item{relHeight}{The relative height of the top summary panel.
Represents the fraction of the plotting area taken up by this panel.}
}
\value{
A \code{ggplot2} object, able to be customised using standard \code{ggplot2} syntax
}
\description{
Plot a Coverage Profile Heatmap across multiple ranges
}
\details{
Convenience function for plotting coverage heatmaps across a common set of
ranges, shared between one or more samples. These are most commonly the
coverage values from merged samples within a treatment group.

A profile data frame here refers to a data.frame (or tibble, or DataFrame)
with a coverage value in one column that corresponds to a genomic bin of a
fixed size, as generated by \link{getProfileData}. Given that multiple
ranges are most likely to be drawn, each profile data frame must be the same
size in terms of the number of bins, each of which represent a fixed number
of nucleotides.
}
