% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSplitDonut.R
\name{plotSplitDonut}
\alias{plotSplitDonut}
\alias{plotSplitDonut,GRanges-method}
\alias{plotSplitDonut,DataFrame-method}
\alias{plotSplitDonut,data.frame-method}
\title{Draw Two-Level Donut Charts}
\usage{
plotSplitDonut(object, ...)

\S4method{plotSplitDonut}{GRanges}(object, scale_by = c("n", "width"), ...)

\S4method{plotSplitDonut}{DataFrame}(object, ...)

\S4method{plotSplitDonut}{data.frame}(
  object,
  inner,
  outer,
  scale_by = NULL,
  r_centre = 0.5,
  r_inner = 1,
  r_outer = 1,
  total_size = 5,
  show_cat = TRUE,
  show_n = FALSE,
  show_percent = TRUE,
  cat_inner = TRUE,
  cat_outer = TRUE,
  label_width = 10,
  sep = " ",
  label_alpha = 1,
  label_size = 4,
  min_p = 0.05,
  explode_inner = NULL,
  explode_outer = NULL,
  explode_query = c("AND", "OR"),
  explode_x = 0,
  explode_y = 0,
  explode_r = 0,
  nudge_r = 0.5,
  expand = 0.1,
  inner_palette = NULL,
  outer_palette = NULL,
  layout = c(main = area(1, 1, 6, 6), lg1 = area(2, 7), lg2 = area(3, 7)),
  ...
)
}
\arguments{
\item{object}{A \code{GRanges} or \code{data.frame}-like object}

\item{...}{Not used}

\item{scale_by}{Column to scale values by. If provided, values in this column
will be summed, instead of simply counting entries. Any label in the centre
of the plot will also reflect this difference}

\item{inner}{Column name to create the inner ring}

\item{outer}{Column name to create the outer ring, subset by the inner ring}

\item{r_centre}{The radius of the hole in the centre. Setting to zero will
create a Pie chart}

\item{r_inner, r_outer}{The radii of the inner/outer rings}

\item{total_size}{Label size total number of entries in the centre of the
plot. Set to \code{NA} to hide the label itself}

\item{show_cat}{logical(1) Show category names for each segment}

\item{show_n}{logical(1) Show category totals for each segment}

\item{show_percent}{logical(1) Show percentages represented by each segment}

\item{cat_inner}{logical(1) Add the column name to labels in the inner ring}

\item{cat_outer}{logical(1) Add the column name to labels in the outer ring}

\item{label_width}{Width at which labels will try to wrap onto a new line}

\item{sep}{character string to be used when concatenating categories, totals
and percentages}

\item{label_alpha}{transparency for labels in the inner ring only}

\item{label_size}{Size of all text labels}

\item{min_p}{only display labels for segments representing greater than this
proportion of the total}

\item{explode_inner, explode_outer}{Regular expressions from either the inner
or outer ring for which segments will be 'exploded'}

\item{explode_query}{Setting to AND and specifying values for both the inner
and outer ring will require matches in both categories}

\item{explode_x, explode_y}{Numeric values for shifting exploded values}

\item{explode_r}{Radius expansion for exploded values}

\item{nudge_r}{Radius expansion for labels in the outer ring}

\item{expand}{Passed to \link[ggplot2]{expansion} for both x and y axes}

\item{inner_palette}{Colour palette for the inner ring}

\item{outer_palette}{Optional colour palette for the outer ring}

\item{layout}{Passed to \link[patchwork]{plot_layout}}
}
\value{
A patchwork object consisting of both ggplot2 objects and legend grobs
}
\description{
Create Donut charts based on one or two columns in a data frame
}
\details{
Using a data.frame or GRanges object, this function enables creation of a
Pie/Donut chart with an inner and outer ring
}
\examples{
set.seed(200)
df <- data.frame(
  feature = sample(
    c("Promoter", "Enhancer", "Intergenic"), 200, replace = TRUE
  ),
  TF1 = sample(c("Up", "Down", "Unchanged"), 200, replace = TRUE),
  TF2 = sample(c("Up", "Down", "Unchanged"), 200, replace = TRUE)
)
## The standard plot
plotSplitDonut(df, inner = "TF1", outer = "TF2")

## Adding an exploded section along with an outer palette & customisation
plotSplitDonut(
  df, inner = "TF1", outer = "feature", total_size = NA,
  label_width = 1, show_n = FALSE, label_alpha = 0.5, r_centre = 0,
  cat_outer = FALSE, explode_inner = "Up", explode_outer = "Prom|Enh",
  explode_query = "AND", explode_r = 0.4,
  inner_palette = hcl.colors(3, "Spectral", rev = TRUE),
  outer_palette = hcl.colors(3, "Cividis")
)

}
